refactor: RankingSection 리팩토링
