+++ README.md
@@ -71,3 +71,8 @@
 ### TYPE 정의 수정
 
 - MAP을 통해 타입을 정의하는 방식으로 변경
+
+### RankingSection 리팩토링
+
+- 컴포넌트 분리
+- 타입 파일 사용
