--- src/features/Home/components/RankingSection.tsx
@@ -1,280 +1,0 @@
-import LoadingSpinner from '@components/common/LoadingSpinner';
-import styled from '@emotion/styled';
-import useFetch from '@hooks/useFetch';
-import { useState } from 'react';
-import { useNavigate, useSearchParams } from 'react-router-dom';
-
-const Target_MAP = {
-  전체: 'ALL',
-  여성이: 'FEMALE',
-  남성이: 'MALE',
-  청소년이: 'TEEN',
-} as const;
-
-// 탭 (랭킹 타입)
-const Rank_MAP = {
-  '받고 싶어한': 'MANY_WISH',
-  '많이 선물한': 'MANY_RECEIVE',
-  '위시로 받은': 'MANY_WISH_RECEIVE',
-} as const;
-
-type TargetType = keyof typeof Target_MAP;
-type RankType = keyof typeof Rank_MAP;
-
-const TargetS = Object.keys(Target_MAP) as TargetType[];
-const RankS = Object.keys(Rank_MAP) as RankType[];
-
-interface Product {
-  id: number;
-  name: string;
-  imageURL: string;
-  price: {
-    basicPrice: number;
-    discountRate: number;
-    sellingPrice: number;
-  };
-  brandInfo: {
-    id: number;
-    name: string;
-    imageURL: string;
-  };
-}
-
-interface RankedProduct extends Product {
-  ranking: number;
-}
-
-const addRanking = (products: Product[]): RankedProduct[] => {
-  return products.map((product, i) => ({
-    ...product,
-    ranking: i + 1,
-  }));
-};
-
-const RankingSection = () => {
-  const [searchParams, setSearchParams] = useSearchParams();
-  const [isExpanded, setIsExpanded] = useState(false);
-
-  // 기본값 설정
-  const rawTarget = searchParams.get('Target');
-  const rawRank = searchParams.get('Rank');
-
-  const selectedTarget: TargetType = TargetS.includes(rawTarget as TargetType)
-    ? (rawTarget as TargetType)
-    : '전체';
-  const selectedRank: RankType = RankS.includes(rawRank as RankType)
-    ? (rawRank as RankType)
-    : '받고 싶어한';
-
-  const updateParam = (key: string, value: string) => {
-    const newParams = new URLSearchParams(searchParams.toString());
-    newParams.set(key, value);
-    setSearchParams(newParams);
-  };
-
-  const { data, loading, hasError } = useFetch<Product[]>(
-    `/products/ranking?targetType=${Target_MAP[selectedTarget]}&rankType=${Rank_MAP[selectedRank]}`
-  );
-
-  const products = data ? addRanking(data) : [];
-
-  //더보기 버튼 로직
-  const visibleItems = isExpanded ? products : products.slice(0, 6);
-
-  const navigate = useNavigate();
-  const handleClick = (item: RankedProduct) => {
-    navigate(`/order/${item.id}`);
-  };
-
-  return (
-    <Section>
-      <Title>실시간 급상승 선물랭킹</Title>
-
-      {/* 필터 */}
-      <TargetRow>
-        {TargetS.map((label) => (
-          <TargetButton
-            key={label}
-            active={selectedTarget === label}
-            onClick={() => updateParam('Target', label)}
-          >
-            {label}
-          </TargetButton>
-        ))}
-      </TargetRow>
-
-      {/* 탭 */}
-      <RankRow>
-        {RankS.map((label) => (
-          <RankButton
-            key={label}
-            active={selectedRank === label}
-            onClick={() => updateParam('Rank', label)}
-          >
-            {label}
-          </RankButton>
-        ))}
-      </RankRow>
-
-      {/* 상품 카드 */}
-      {loading && <LoadingSpinner />}
-
-      {!loading && !hasError && products.length > 0 ? (
-        <Grid>
-          {visibleItems.map((item: RankedProduct) => (
-            <Card key={item.ranking}>
-              {/* 임시로 ranking으로 해두었지만 추후 id값으로 바꿀 계획 */}
-              <ImageWrapper onClick={() => handleClick(item)}>
-                <ProductImage src={item.imageURL} alt={item.name} />
-                <RankBadge>{item.ranking}</RankBadge>
-              </ImageWrapper>
-              <Brand>{item.brandInfo.name}</Brand>
-              <ProductName>{item.name}</ProductName>
-              <Price>{item.price.sellingPrice.toLocaleString()} 원</Price>
-            </Card>
-          ))}
-        </Grid>
-      ) : (
-        <EmptyMessage>상품이 없습니다.</EmptyMessage>
-      )}
-
-      {/* 더보기 / 접기 버튼 */}
-      <ToggleButton onClick={() => setIsExpanded((prev) => !prev)}>
-        {isExpanded ? '접기' : '더보기'}
-      </ToggleButton>
-    </Section>
-  );
-};
-
-export default RankingSection;
-
-const Section = styled.section(({ theme }) => ({
-  padding: theme.spacing.spacing4,
-}));
-
-const Title = styled.h2(({ theme }) => ({
-  fontSize: theme.typography.title2Bold.fontSize,
-  fontWeight: theme.typography.title2Bold.fontWeight,
-  color: theme.colors.semantic.textDefault,
-  marginBottom: theme.spacing.spacing4,
-}));
-
-const TargetRow = styled.div(({ theme }) => ({
-  display: 'flex',
-  justifyContent: 'space-between',
-  gap: theme.spacing.spacing2,
-  marginBottom: theme.spacing.spacing3,
-  padding: theme.spacing.spacing2,
-}));
-
-const TargetButton = styled.button<{ active: boolean }>(
-  ({ theme, active }) => ({
-    padding: `${theme.spacing.spacing2} ${theme.spacing.spacing3}`,
-    borderRadius: '20px',
-    backgroundColor: active
-      ? theme.colors.blue.blue600
-      : theme.colors.gray.gray100,
-    color: active ? '#fff' : theme.colors.gray.gray700,
-    border: 'none',
-    fontSize: theme.typography.label2Regular.fontSize,
-    cursor: 'pointer',
-    fontWeight: 600,
-  })
-);
-
-const RankRow = styled.div(({ theme }) => ({
-  display: 'flex',
-  justifyContent: 'space-between',
-  backgroundColor: theme.colors.gray.gray100,
-  padding: theme.spacing.spacing1,
-  borderRadius: '8px',
-  marginBottom: theme.spacing.spacing4,
-}));
-
-const RankButton = styled.button<{ active?: boolean }>(({ theme, active }) => ({
-  flex: 1,
-  padding: theme.spacing.spacing2,
-  fontSize: theme.typography.body2Bold.fontSize,
-  fontWeight: theme.typography.body2Bold.fontWeight,
-  border: 'none',
-  borderRadius: '6px',
-  backgroundColor: active ? theme.colors.blue.blue600 : 'transparent',
-  color: active ? '#fff' : theme.colors.gray.gray600,
-  cursor: 'pointer',
-  transition: 'background-color 0.2s ease',
-}));
-
-const Grid = styled.div(({ theme }) => ({
-  display: 'grid',
-  gridTemplateColumns: 'repeat(3, 1fr)',
-  gap: theme.spacing.spacing4,
-}));
-
-const Card = styled.div`
-  text-align: center;
-`;
-
-const ImageWrapper = styled.div`
-  position: relative;
-  width: 100%;
-  aspect-ratio: 1/1;
-  cursor: pointer;
-`;
-
-const ProductImage = styled.img`
-  width: 100%;
-  height: auto;
-  border-radius: 8px;
-`;
-
-const RankBadge = styled.div(({ theme }) => ({
-  position: 'absolute',
-  top: '6px',
-  left: '6px',
-  backgroundColor: theme.colors.red.red700,
-  color: '#fff',
-  padding: '2px 6px',
-  borderRadius: '12px',
-  fontSize: '12px',
-  fontWeight: 'bold',
-}));
-
-const Brand = styled.p(({ theme }) => ({
-  marginTop: theme.spacing.spacing2,
-  fontSize: theme.typography.label2Regular.fontSize,
-  color: theme.colors.gray.gray600,
-}));
-
-const ProductName = styled.p(({ theme }) => ({
-  fontWeight: theme.typography.body2Bold.fontWeight,
-  fontSize: theme.typography.body2Bold.fontSize,
-  margin: 0,
-}));
-
-const Price = styled.p(({ theme }) => ({
-  marginTop: '4px',
-  fontWeight: theme.typography.body2Bold.fontWeight,
-  fontSize: theme.typography.body2Bold.fontSize,
-}));
-
-const EmptyMessage = styled.div(({ theme }) => ({
-  ...theme.typography.body1Regular,
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  minHeight: '28.75rem',
-}));
-
-const ToggleButton = styled.button(({ theme }) => ({
-  display: 'block',
-  margin: `${theme.spacing.spacing4} auto 0`,
-  marginTop: theme.spacing.spacing4,
-  padding: `${theme.spacing.spacing3} ${theme.spacing.spacing10}`,
-  border: `1px solid ${theme.colors.gray.gray300}`,
-  background: 'white',
-  color: theme.colors.gray.gray800,
-  borderRadius: '8px',
-  fontSize: theme.typography.body2Regular.fontSize,
-  cursor: 'pointer',
-  fontWeight: 500,
-}));
