+++ src/features/Home/components/ranking/ProductGrid.tsx
@@ -1,0 +1,131 @@
+import LoadingSpinner from '@components/common/LoadingSpinner';
+import type { GridProps, RankedProduct } from './RankingTypes';
+import styled from '@emotion/styled';
+
+const ProductGrid = ({
+  products,
+  loading,
+  hasError,
+  isExpanded,
+  toggleExpand,
+  onClickItem,
+}: GridProps) => {
+  const visibleItems = isExpanded ? products : products.slice(0, 6);
+  return (
+    <>
+      {loading && <LoadingSpinner />}
+
+      {!loading && !hasError && products.length > 0 ? (
+        <Grid>
+          {visibleItems.map((item: RankedProduct) => (
+            <ProductCard item={item} onClickItem={onClickItem} />
+          ))}
+        </Grid>
+      ) : (
+        <EmptyMessage>상품이 없습니다.</EmptyMessage>
+      )}
+
+      {/* 더보기 / 접기 버튼 */}
+      <ToggleButton onClick={toggleExpand}>
+        {isExpanded ? '접기' : '더보기'}
+      </ToggleButton>
+    </>
+  );
+};
+
+export default ProductGrid;
+
+interface CardProps {
+  item: RankedProduct;
+  onClickItem: (item: RankedProduct) => void;
+}
+const ProductCard = ({ item, onClickItem }: CardProps) => {
+  return (
+    <Card key={item.ranking}>
+      {/* 임시로 ranking으로 해두었지만 추후 id값으로 바꿀 계획입니다 */}
+      {/*Card도 따로 파일을 분리하려 했으나 오히려 더 가독성이 나빠질 수도 있을 것 같아 보류했습니다.*/}
+      <ImageWrapper onClick={() => onClickItem(item)}>
+        <ProductImage src={item.imageURL} alt={item.name} />
+        <RankBadge>{item.ranking}</RankBadge>
+      </ImageWrapper>
+      <Brand>{item.brandInfo.name}</Brand>
+      <ProductName>{item.name}</ProductName>
+      <Price>{item.price.sellingPrice.toLocaleString()} 원</Price>
+    </Card>
+  );
+};
+
+const Grid = styled.div(({ theme }) => ({
+  display: 'grid',
+  gridTemplateColumns: 'repeat(3, 1fr)',
+  gap: theme.spacing.spacing4,
+}));
+
+const Card = styled.div`
+  text-align: center;
+`;
+
+const ImageWrapper = styled.div`
+  position: relative;
+  width: 100%;
+  aspect-ratio: 1/1;
+  cursor: pointer;
+`;
+
+const ProductImage = styled.img`
+  width: 100%;
+  height: auto;
+  border-radius: 8px;
+`;
+
+const RankBadge = styled.div(({ theme }) => ({
+  position: 'absolute',
+  top: '6px',
+  left: '6px',
+  backgroundColor: theme.colors.red.red700,
+  color: '#fff',
+  padding: '2px 6px',
+  borderRadius: '12px',
+  fontSize: '12px',
+  fontWeight: 'bold',
+}));
+
+const Brand = styled.p(({ theme }) => ({
+  marginTop: theme.spacing.spacing2,
+  fontSize: theme.typography.label2Regular.fontSize,
+  color: theme.colors.gray.gray600,
+}));
+
+const ProductName = styled.p(({ theme }) => ({
+  fontWeight: theme.typography.body2Bold.fontWeight,
+  fontSize: theme.typography.body2Bold.fontSize,
+  margin: 0,
+}));
+
+const Price = styled.p(({ theme }) => ({
+  marginTop: '4px',
+  fontWeight: theme.typography.body2Bold.fontWeight,
+  fontSize: theme.typography.body2Bold.fontSize,
+}));
+
+const EmptyMessage = styled.div(({ theme }) => ({
+  ...theme.typography.body1Regular,
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  minHeight: '28.75rem',
+}));
+
+const ToggleButton = styled.button(({ theme }) => ({
+  display: 'block',
+  margin: `${theme.spacing.spacing4} auto 0`,
+  marginTop: theme.spacing.spacing4,
+  padding: `${theme.spacing.spacing3} ${theme.spacing.spacing10}`,
+  border: `1px solid ${theme.colors.gray.gray300}`,
+  background: 'white',
+  color: theme.colors.gray.gray800,
+  borderRadius: '8px',
+  fontSize: theme.typography.body2Regular.fontSize,
+  cursor: 'pointer',
+  fontWeight: 500,
+}));
